packages/graph/pyproject.toml
[project]
name = "graph"
version = "0.1.0"
description = "Neo4j graph schema and loader for Weed WizAIrd herbicide knowledge graph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graph"]

[project.scripts]
load-graph = "graph.loader:main"
init-schema = "graph.schema:main"
